% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhhrvrditemserver.R
\name{addrditemserver}
\alias{addrditemserver}
\title{Title 新增数据}
\usage{
addrditemserver(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
Title 新增数据
}
\examples{
addrditemserver()  
}
